Integrate formal charge and bond order into CherryPicker:
CherryPicker and this repo use the same namespace, indigox, and have some classes that share names but have different functions between the repos, e.g. atom, molecule. This has made it hard to integrate this repo into CherryPicker as a submodule. Here I've attempted to change the namespace of this repo to avoid the clashes.